+++ src/main/java/absolutelyaya/ultracraft/client/rendering/entity/other/InterruptableChargeRenderer.java
@@ -1,7 +1,6 @@
 package absolutelyaya.ultracraft.client.rendering.entity.other;
 
 import absolutelyaya.ultracraft.Ultracraft;
-import absolutelyaya.ultracraft.client.UltracraftClient;
 import absolutelyaya.ultracraft.entity.other.InterruptableCharge;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.render.*;
@@ -9,15 +8,13 @@ import net.minecraft.client.render.entity.EntityRenderer;
 import net.minecraft.client.render.entity.EntityRendererFactory;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.util.Identifier;
+import org.joml.Matrix4f;
 
 public class InterruptableChargeRenderer extends EntityRenderer<InterruptableCharge>
 {
-	final InterruptableChargeModel model;
-	
 	public InterruptableChargeRenderer(EntityRendererFactory.Context ctx)
 	{
 		super(ctx);
-		model = new InterruptableChargeModel(ctx.getModelLoader().getModelPart(UltracraftClient.INTERRUPTABLE_CHARGE_LAYER));
 	}
 	
 	@Override
@@ -32,10 +29,15 @@ public class InterruptableChargeRenderer extends EntityRenderer<InterruptableCha
 		RenderLayer renderLayer = RenderLayer.getEntityTranslucent(getTexture(entity));
 		VertexConsumer vertexConsumer = vertexConsumers.getBuffer(renderLayer);
 		matrices.push();
-		float f = entity.getScale();
+		float f = entity.getScale() / 2f;
 		matrices.scale(f, f, f);
 		matrices.multiply(MinecraftClient.getInstance().gameRenderer.getCamera().getRotation());
-		model.render(matrices, vertexConsumer, light, OverlayTexture.DEFAULT_UV, 1f, 1f, 1f, 1f);
+		matrices.translate(0f, 0.25f, 0f);
+		Matrix4f matrix = matrices.peek().getPositionMatrix();
+		vertexConsumer.vertex(matrix, -0.5f, -0.5f, 0).color(1f, 1f, 1f, 1f).texture(0f, 0f).overlay(OverlayTexture.DEFAULT_UV).light(LightmapTextureManager.MAX_BLOCK_LIGHT_COORDINATE).normal(0f, 1f, 0f).next();
+		vertexConsumer.vertex(matrix, -0.5f, 0.5f, 0).color(1f, 1f, 1f, 1f).texture(0f, 0.5f).overlay(OverlayTexture.DEFAULT_UV).light(LightmapTextureManager.MAX_BLOCK_LIGHT_COORDINATE).normal(0f, 1f, 0f).next();
+		vertexConsumer.vertex(matrix, 0.5f, 0.5f, 0).color(1f, 1f, 1f, 1f).texture(0.5f, 0.5f).overlay(OverlayTexture.DEFAULT_UV).light(LightmapTextureManager.MAX_BLOCK_LIGHT_COORDINATE).normal(0f, 1f, 0f).next();
+		vertexConsumer.vertex(matrix, 0.5f, -0.5f, 0).color(1f, 1f, 1f, 1f).texture(0.5f, 0f).overlay(OverlayTexture.DEFAULT_UV).light(LightmapTextureManager.MAX_BLOCK_LIGHT_COORDINATE).normal(0f, 1f, 0f).next();
 		matrices.pop();
 	}